requirements.txt
fastapi
python-dotenv
google-cloud-storage
uvicorn
aiomysql